+++ pyspark/bigdl/keras1/backend.py
@@ -1,9 +1,7 @@
-import new
-import types # TODO: support python3
+# import new
+# import types # TODO: support python3
 import tempfile
 
-from bigdl.keras1.converter import BigDLLayerConverter
-from bigdl.util.common import callBigDlFuncWithoutMappingReturn
 from bigdl.util.common import get_spark_context
 from bigdl.nn.layer import Model as BModel
 import bigdl.nn.layer as BLayer
@@ -18,11 +16,6 @@ from keras.models import model_from_json
 # > x.method = types.MethodType(method, x)
 from keras.models import Sequential, Model
 import numpy as np
-
-def create_tmp_path():
-    tmp_file = tempfile.NamedTemporaryFile(prefix="bigdl")
-    tmp_file.close()
-    return tmp_file.name
 
 
 def install_bigdl_backend(kmodel):
@@ -72,230 +65,4 @@ def install_bigdl_backend(kmodel):
     kmodel.predict = bpredict
     kmodel.evaluate = bevaluate
     kmodel.get_weights = bmodel.get_weights
-
-class ModelLoader:
-
-    @staticmethod
-    def to_bigdl_criterion(kloss):
-        # TODO: it may pass in an object
-        if kloss == "categorical_crossentropy":
-            return bcriterion.ClassNLLCriterion()
-        elif kloss == "mse":
-            return bcriterion.MSECriterion()
-        else:
-            raise Exception("Not supported type: %s" % kloss)
-
-    @staticmethod
-    def to_bigdl_optim_method(koptim_method):
-        # This is always be an object
-        if isinstance(koptim_method, koptimizers.Adagrad):
-            return boptimizer.Adagrad()
-        elif isinstance(koptim_method, koptimizers.SGD):
-            return boptimizer.SGD(learningrate=0.01)  # TODO: enrich parameters, sgd.lr return a variable!!!not float
-        else:
-            raise Exception("Not supported type: %s" % koptim_method)
-
-    @staticmethod
-    def to_sample_rdd(sc, x, y):
-        from bigdl.util.common import Sample
-        x_rdd = sc.parallelize(x)
-        y_rdd = sc.parallelize(y)
-        return x_rdd.zip(y_rdd).map(lambda item: Sample.from_ndarray(item[0], item[1]))
-
-    @staticmethod
-    def to_bigdl_definition(kmodel):
-        keras_model_path = create_tmp_path()
-        keras_model_path_json = keras_model_path + ".json"
-        #keras_model_path_hdf5 = keras_model_path + ".hdf5"
-
-        with open(keras_model_path_json, "w") as json_file:
-            json_file.write(kmodel.to_json())
-        #kmodel.save(keras_model_path_hdf5)
-
-        # load bigdl model from file. TODO: no need to save json to file first
-        return ModelLoader.load_definition(keras_model_path_json)
-
-    @staticmethod
-    def load_definition(model_path, bigdl_type="float"):
-        bigdl_graph = callBigDlFuncWithoutMappingReturn(bigdl_type,
-                                                        "kerasLoadGraph", model_path)
-        return BModel.from_jvalue(bigdl_graph)
-
-    # this code is from keras
-    # model is a BModel
-    @staticmethod
-    def load_weights(bmodel, kmodel, filepath, by_name=False):
-        '''Loads all layer weights from a HDF5 save file.
-
-        If `by_name` is False (default) weights are loaded
-        based on the network's topology, meaning the architecture
-        should be the same as when the weights were saved.
-        Note that layers that don't have weights are not taken
-        into account in the topological ordering, so adding or
-        removing layers is fine as long as they don't have weights.
-
-        If `by_name` is True, weights are loaded into layers
-        only if they share the same name. This is useful
-        for fine-tuning or transfer-learning models where
-        some of the layers have changed.
-        '''
-        import h5py
-        f = h5py.File(filepath, mode='r')
-        if 'layer_names' not in f.attrs and 'model_weights' in f:
-            f = f['model_weights']
-        else:
-            raise Exception("not supported format")
-        if by_name:
-            pass  # load_weights_from_hdf5_group_by_name(f)
-        else:
-            ModelLoader.load_weights_from_hdf5_strict(f, bmodel, kmodel)
-        if hasattr(f, 'close'):
-            f.close()
-
-    @staticmethod
-    def keras_name_to_Layer(kmodel):
-        layers = kmodel.layers
-        return dict([(layer.name, layer) for layer in layers])
-
-    @staticmethod
-    def load_weights_from_hdf5_strict(f, bmodel, kmodel):
-        '''Weight loading is based on layer order in a list
-        (matching model.flattened_layers for Sequential models,
-        and model.layers for Model class instances), not
-        on layer names.
-        Layers that have no weights are skipped.
-        '''
-        keras_name_to_layer = ModelLoader.keras_name_to_Layer(kmodel)
-
-        if 'nb_layers' in f.attrs:
-            raise Exception("not supported format of old version")
-        else:
-            blayers = bmodel.executions()
-            bigdl_layers_wb = [layer for layer in blayers if layer.get_weights()]
-
-            layer_names = [n.decode('utf8') for n in f.attrs['layer_names']]
-            filtered_layer_names = []
-            for name in layer_names:
-                g = f[name]
-                weight_names = [n.decode('utf8') for n in g.attrs['weight_names']]
-                if len(weight_names):
-                    filtered_layer_names.append(name)
-            if len(filtered_layer_names) != len(bigdl_layers_wb):
-                raise Exception('You are trying to load a weight file '
-                                'containing ' + str(len(layer_names)) +
-                                ' layers into a model with ' +
-                                str(len(bigdl_layers_wb)) + ' layers.')
-            # ensure the layers between bigdl and keras are identical
-            bigdl_layers_wb = sorted(bigdl_layers_wb, key=lambda layer: layer.name())
-            filtered_layer_names = sorted(filtered_layer_names)
-            for bwlayer, klayer in zip():
-                if bwlayer != klayer:
-                    raise Exception(
-                        "Found bigdl layer %s, keras layer: %s" % (bwlayer.name(), klayer))
-
-            for k, name in enumerate(filtered_layer_names):
-                g = f[name]
-                weight_names = [n.decode('utf8') for n in g.attrs['weight_names']]
-                # TODO: performance?
-                weight_values = [np.array(g[weight_name]) for weight_name in weight_names]
-                bigdl_layer = bigdl_layers_wb[k]
-                # TODO: Check if the len of weights is equal
-                # symbolic_weights = layer.weights
-                # if len(weight_values) != len(symbolic_weights):
-                #     raise Exception('Layer #' + str(k) +
-                #                     ' (named "' + layer.name +
-                #                     '" in the current model) was found to '
-                #                     'correspond to layer ' + name +
-                #                     ' in the save file. '
-                #                     'However the new layer ' + layer.name +
-                #                     ' expects ' + str(len(symbolic_weights)) +
-                #                     ' weights, but the saved weights have ' +
-                #                     str(len(weight_values)) +
-                #                     ' elements.')
-                if layer.__class__.__name__ == 'Convolution1D':
-                    # This is for backwards compatibility with
-                    # the old Conv1D weights format.
-                    raise Exception("we don't support old format for now")
-                bigdl_weights = ParameterConverter.convert_weights(
-                        keras_name_to_layer[name].__class__.__name__,
-                                                                   weight_values)
-                bigdl_layer.set_weights(bigdl_weights)
-
-class ParameterConverter:
-    @staticmethod
-    def convert_weights(layer_name, wb):
-        if layer_name == "Dense":
-            return [np.transpose(wb[0]), wb[1]]
-        else:
-            return wb
-
-
-class DefinitionLoader:
-
-    def __init__(self, json_path):
-        self.node_id_to_instance = {}
-        self.node_id_to_layer = {}
-        self.node_id_to_config_layer = {}
-        with open(json_path, "r") as jp:
-            self.kmodel = model_from_json(jp.read())
-            self.kconfig = self.kmodel.get_config()
-
-        for layer in self.kmodel.layers:
-            self.node_id_to_layer[layer.name] = layer
-
-        for clayer in self.kconfig["layers"]:
-            self.node_id_to_config_layer[clayer["name"]] = clayer
-
-    def _create_bigdl_layer(self, class_name, layer):
-        pass
-    def _do_create_node(self, layer, clayer):
-        if clayer["class_name"] == "InputLayer":
-            input = BLayer.Input()
-            input.element().set_name(layer.name) # cannot set name for node?
-            self.node_id_to_instance[layer.name] = input
-            return input
-        bigdl_in_nodes = []
-        for node in clayer["inbound_nodes"]:
-            for out in node:
-                out_name = out[0]
-                out_index = out[1]
-                out_tensor_index = out[2]
-                if out_name not in self.node_id_to_instance:
-                    self._do_create_node(self.node_id_to_layer[out_name],
-                                         self.node_id_to_config_layer[out_name])
-                bigdl_in_nodes.append(self.node_id_to_instance[out_name])
-
-        blayer = BigDLLayerConverter().create(clayer["class_name"], layer, clayer)
-        new_bnode = blayer(bigdl_in_nodes)
-        self.node_id_to_instance[layer.name] = new_bnode
-        return new_bnode
-
-    def _construct_bigdl_model(self):
-        for clayer in self.kconfig["layers"]:
-            if clayer["name"] not in self.node_id_to_instance:
-
-                self._do_create_node(self.node_id_to_layer[clayer["name"]],
-                                     clayer)
-        ins = []
-        for input_layer in self.kconfig["input_layers"]:
-            name = input_layer[0]
-            ins.append(self.node_id_to_instance[name])
-        outs = []
-        for output_layer in self.kconfig["output_layers"]:
-            name = output_layer[0]
-            outs.append(self.node_id_to_instance[name])
-        return BLayer.Model(inputs=ins, outputs=outs)
-
-    def _construct_bigdl_sequence(self):
-        pass
-
-
-    def to_bigdl(self):
-        if isinstance(self.kmodel, Model):
-            bmodel = self._construct_bigdl_model()
-        elif isinstance(self.kmodel, Sequential):
-            bmodel = self. _construct_bigdl_sequence()
-        return bmodel
-
-
