+++ pyspark/test/bigdl/keras1/test_load_definition.py
@@ -40,7 +40,7 @@ from keras.utils import np_utils
 class TestLoadDefinition():
 
     def test_load_mlp(self):
-        loader = bigdl_backend.DefinitionLoader("/Users/lizhichao/god/BigDL/pyspark/test/bigdl/keras1/multiple_inputs_model.json")
+        loader = bigdl_backend.DefinitionLoader.from_json("/Users/lizhichao/god/BigDL/pyspark/test/bigdl/keras1/multiple_inputs_model.json")
         loader.to_bigdl()
 
 if __name__ == "__main__":
